refactor(ListMatkulIcon): move ListMatkulIcon to asset folder
